Added auto shift and OTA capabilities
Added pins ecuAutoUpShift, ecuLanch, enableAutoUpShift  with temporary numbers(6,7,8). Over the air(OTA)  code upload is available, make sure you always leave that section in other wise we will need to remove the module and upload by USB.  The password to uploading in 4992 at the moment. It is using InteralStorage (not SD) for OTA so only programs that are half of the memory of the MKR can be uploaded.